Created a utils file with functions for testing.
Gets rid of a lot of the error prone and repetetive
database connections that were being used to make
tiny manipulations to the database, without using
the code that was actually being tested.

Seems redundant to have these functions in general,
but I think it's useful to separate intent for
when the database is manipulated.